# src/main/java/ru/demo/companyServise/company/src/main/java/com/example/company/model/User.java
package ru.demo.companyServise.company.src.main.java.com.example.company.model;


import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class User {
    private Long id;
    private String firstName;
    private String lastName;
    private String phoneNumber;
    private String email;
    private Long companyId;


}