# src/main/java/ru/demo/userService/userService/src/main/java/com/example/userService/dto/UserUpdateDto.java
package ru.demo.userService.userService.src.main.java.com.example.userService.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotEmpty;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class UserUpdateDto {

    @NotEmpty(message = "First name cannot be empty")
    private String firstName;
    @NotEmpty(message = "Last name cannot be empty")
    private String lastName;
    @NotEmpty(message = "phoneNumber cannot be empty")
    private String phoneNumber;
    @Email(message = "Email should be valid")
    private String email;
}